Fix run-to-run nondeterminism and sticky fast-forward
The render-interpolation save buffers were Float32Array: every rendered
frame between physics steps truncated the f64 sim state, so trajectories
depended on display refresh rate and frame timing — same track, visibly
different ride each play (~83px divergence after 10s in the new smoke
check). Buffers are now Float64Array, making the lerp round-trip
bit-exact; a permanent 'lerp determinism' smoke test guards it.

Also reset playback speed to 1x on play/restart — a sticky 2x from a
previous run made new runs feel randomly faster.

+++ src/physics/rider.ts
@@ -36,9 +36,12 @@ export class Rider {
   readonly scarf: ScarfNode[] = [];
   crashed = false;
 
-  /** Positions at the start of the last step — the render-interpolation baseline. */
-  private prevState!: Float32Array;
-  private lerpSave!: Float32Array;
+  /** Positions at the start of the last step — the render-interpolation baseline.
+   *  MUST be Float64: state round-trips through these buffers every rendered
+   *  frame, and truncating to f32 would perturb the sim differently per
+   *  display refresh rate, breaking run-to-run determinism. */
+  private prevState!: Float64Array;
+  private lerpSave!: Float64Array;
   private lerping = false;
 
   constructor() {
@@ -72,8 +75,8 @@ export class Rider {
       this.scarf.push({ x: 0, y: 0, px: 0, py: 0 });
     }
     const n = (this.allPoints.length + this.scarf.length) * 4;
-    this.prevState = new Float32Array(n);
-    this.lerpSave = new Float32Array(n);
+    this.prevState = new Float64Array(n);
+    this.lerpSave = new Float64Array(n);
     this.reset(0, 0);
   }
 
@@ -97,7 +100,7 @@ export class Rider {
     this.capturePrev();
   }
 
-  private writeState(out: Float32Array): void {
+  private writeState(out: Float64Array): void {
     let i = 0;
     for (const p of this.allPoints) {
       out[i++] = p.x;
